Delete note attempt branch

+++ src/actions/tagActions.js
@@ -3,7 +3,6 @@ import * as types from '../constants/actionTypes';
 
 import { createNewTag, getTagCount } from '../common/noteHelpers';
 import { uniq } from '../common/helpers';
-import { DEFAULTS } from '../constants/noteConst';
 
 export function getTags() {
     return dispatch => {
@@ -37,11 +36,11 @@ export function addTag(tags, note) {
         // Make tag list unique
         tags = uniq(tags);
         let tagList = [];
-        
+
         // Only add new tags but make full tagList
         tags.forEach((tag) => {
             const tagsRef = database.ref('tags');
-            
+
             // if no ID push a new tag to the list
             if (!tag.id && tag.className) {
                 const tagRef = tagsRef.push();
@@ -49,7 +48,7 @@ export function addTag(tags, note) {
 
                 tagRef.set(tag);
             }
-            
+
             // push to tagList
             tagList.push(tag);
         });
@@ -59,39 +58,60 @@ export function addTag(tags, note) {
 }
 
 export function removeTags(notes) {
-	return dispatch => {
+	return (dispatch, getState) => {
 		dispatch(deleteTagsRequestedAction());
 
+        const user = getState().userData.user;
         const tagsRef = database.ref('tags');
 
         tagsRef.once('value', (snap) => {
-            const tags = snap.val();
-            let tagsList = [];
-            
-            Object.keys(tags).forEach((t) => {
-                let tag = tags[t];
-                let tagCount = getTagCount(tag, notes);
+            if (snap.exists()) {
+                const tags = snap.val();
+                let tagsList = [];
 
-                // Remove empty tags
-                if (tagCount.count === 0 && tagCount.tag.name !== DEFAULTS.TAG) {
-                    let tagRef = tagsRef.child(tagCount.tag.id);
-                    // remove tag
-                    tagRef.remove();
-                } else {
-                    tagsList.push(tags[t]);
-                }
-            });
-            
-            dispatch(deleteTagsFulfilledAction(tagsList));
+                Object.keys(tags).forEach((t) => {
+                    let tag = tags[t];
+                    let tagCount = getTagCount(tag, notes, user);
+
+                    // Remove empty tags
+                    if (tagCount.count === 0) {
+                        let tagRef = tagsRef.child(tagCount.tag.id);
+                        // remove tag
+                        tagRef.remove();
+                    } else {
+                        tagsList.push(tags[t]);
+                    }
+                });
+
+                dispatch(deleteTagsFulfilledAction(tagsList));
+            }
         });
 	};
 }
 
 export function selectTag(tag) {
     return (dispatch, getState) => {
-        
+
         const tag = getState().tagData.tags.filter(function(n) {
             return n.id = tag.id;
+        });
+    }
+}
+
+export function listenForDeletedTags() {
+    return (dispatch, getState) => {
+        const notesRef = database.ref('notes');
+
+        notesRef.on('child_removed', (snap) => {
+            let notes = getState().noteData.notes;
+            const n = snap.val();
+
+            // Filter the deleted note out of current notes state
+            notes = notes.filter((note) => {
+                return note.id !== n.id;
+            });
+
+            dispatch(removeTags(notes));
         });
     }
 }
